resume-processing-service/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.17
pydantic==2.9.2
spacy==3.8.2
PyMuPDF==1.24.13
python-docx==1.1.2
openai==1.54.4
